+++ src/terminal/ParserEvents.h
@@ -150,7 +150,7 @@ class ParserEvents
     virtual void dispatchAPC() = 0;
 
     virtual void startPM() = 0;
-    virtual void putPM(char32_t) = 0;
+    virtual void putPM(char) = 0;
     virtual void dispatchPM() = 0;
 };
 
@@ -177,7 +177,7 @@ class BasicParserEvents: public ParserEvents
     void putAPC(char) override {}
     void dispatchAPC() override {}
     void startPM() override {}
-    void putPM(char32_t) override {}
+    void putPM(char) override {}
     void dispatchPM() override {}
 };
 } // end namespace terminal